Rename & Style: api 파일 이름 변경 및 info 너비 설정
