+++ src/components/StartupDetail/StartupDetailInfo.module.css
@@ -24,7 +24,7 @@
   align-items: flex-start;
   padding: 2.4rem;
   gap: 1rem;
-  width: 119.6rem;
+  width: 100%;
   height: 28.7rem;
   left: 36.4rem;
   top: 36.9rem;
